Organize types/functions into one file.
